requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-cors-headers==2.5.3
django-user-agents==0.3.2
djangorestframework==3.9.2
Faker==1.0.4
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
typed-ast==1.3.1
ua-parser==0.8.0
urllib3==1.24.1
user-agents==2.0
wrapt==1.11.1
